Re/Add (middleware)
- Delete a separate middleware/ module by transferring its contents to appropriate modules (ska\middleware.py, authentication\middleware.py)

- Change connection of transfered middleware items (ska\settings.py)

- Add structural comments
